Isolate dependencies on test suite headers
